Fix environment envoding for snippets
The environment for snippets was not correctly
setup. In particular, not all strings were
encoded in utf-8, leading to problems when
for example filenames were not encoded in utf-8
and used in placeholders.

This commit introduces two separate environments,
one with proper encoded utf-8 and one with the
original encoding of all the various strings.
The original encoding environment is used only
for shell placeholders.