Remove the gedit style scheme manager instance
Instead of reimplementing our singleton for the style scheme manager,
leverage the gtksourceview one, so that plugins can simply obtain it
with get_default(). We just need to make sure the search path is set
at startup before any document is opened.
The utility functions to install/uninstall a style scheme are now in
the pref dialog file since in five years we never needed to export them
or use them in another context.